Move code from preprocessing_tools.py to preprocessor module
